+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.3.1 - 2025-01-03
+
+### Fixed
+
+- Notes editor was broken (when editing a ride). Fixed by upgrading to the fork of react-quill-new.
+
 ## 3.3.0 - 2024-12-28
 
 ### Added